bug_localization/requirements.txt
datasets==2.15.0
huggingface_hub==0.27.1
hydra-core==1.3.2
numpy==1.26.4
tokenizers==0.21.0
transformers==4.48.1
omegaconf==2.3.0
openai==1.60.1
nltk==3.9.1
scikit-learn==1.6.1
gitpython==3.1.44
matplotlib==3.10.0
pandas~=2.2.3
pytest~=8.3.4
langdetect~=1.0.9
tenacity~=8.5.0
python-dotenv~=1.0.1
anthropic~=0.45.0
tiktoken~=0.8.0
unidiff~=0.7.5
torch~=2.5.1
backoff~=2.2.1
python-dotenv~=1.0.1
huggingface-hub~=0.27.1
tree_sitter~=0.20.0